sckgs_najmulhasan/sckgs_backend/requirements.txt
asgiref==3.7.2
Django==5.0
djangorestframework==3.14.0
Faker==21.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
